Patch

+++ plugins/core/nodegraph/src/Madgine/nodegraph/graphexecutable.h
@@ -6,7 +6,7 @@ namespace Engine {
 namespace NodeGraph {
 
 	struct GraphExecutable : Serialize::SerializableUnitBase{
-        virtual void execute(NodeInstance *instance, size_t flowIndex, ArgumentList *out = nullptr) const = 0;
+        virtual void execute(NodeInstance *instance, uint32_t flowIndex, ArgumentList *out = nullptr) const = 0;
 	};
 
 }
