+++ plugins/core/nodegraph/src/Madgine/nodegraph/nodes/testnode.h
@@ -11,16 +11,16 @@ namespace NodeGraph {
         TestNode();
 
         virtual size_t flowInCount() const override;
-        virtual std::string_view flowInName(size_t index) const override;
+        virtual std::string_view flowInName(uint32_t index) const override;
 
-        virtual std::string_view flowOutName(size_t index) const override;
+        virtual std::string_view flowOutName(uint32_t index) const override;
 
-        virtual std::string_view dataInName(size_t index) const override;
-        virtual ExtendedValueTypeDesc dataInExpectedType(size_t index) const override;
+        virtual std::string_view dataInName(uint32_t index) const override;
+        virtual ExtendedValueTypeDesc dataInExpectedType(uint32_t index) const override;
 
-        virtual std::string_view dataOutName(size_t index) const override;
+        virtual std::string_view dataOutName(uint32_t index) const override;
 
-        virtual void execute(NodeInstance *instance, size_t flowIndex, ArgumentList *out = nullptr) const override;
+        virtual void execute(NodeInstance *instance, uint32_t flowIndex, ArgumentList *out = nullptr) const override;
     };
 
 }
